Base types are moved to API subproject.
