+++ UnityProject/Assets/Zenject/Main/Scripts/Factories/Factory.cs
@@ -2,6 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Security.Permissions;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+using ModestTree.Zenject.Api.Factories;
 
 namespace ModestTree.Zenject
 {
