+++ UnityProject/Assets/Zenject/Main/Scripts/Factories/FactoryMethod.cs
@@ -1,5 +1,8 @@
 using System;
 using System.Collections.Generic;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+using ModestTree.Zenject.Api.Factories;
 
 namespace ModestTree.Zenject
 {
