+++ UnityProject/Assets/Zenject/Main/Scripts/Factories/FactoryTyped.cs
@@ -1,19 +1,11 @@
 using System;
 using System.Collections.Generic;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+using ModestTree.Zenject.Api.Factories;
 
 namespace ModestTree.Zenject
 {
-    public interface IValidatable
-    {
-        IEnumerable<ZenjectResolveException> Validate();
-    }
-
-    // Zero parameters
-    public interface IFactoryTyped<TValue> : IValidatable
-    {
-        TValue Create();
-    }
-
     public class FactoryTyped<TValue> : IFactoryTyped<TValue>
     {
         [Inject]
@@ -28,12 +20,6 @@ namespace ModestTree.Zenject
         {
             return _container.ValidateObjectGraph<TValue>();
         }
-    }
-
-    // One parameter
-    public interface IFactoryTyped<TParam1, TValue> : IValidatable
-    {
-        TValue Create(TParam1 param);
     }
 
     public class FactoryTyped<TParam1, TValue>
@@ -53,12 +39,6 @@ namespace ModestTree.Zenject
         }
     }
 
-    // Two parameters
-    public interface IFactoryTyped<TParam1, TParam2, TValue> : IValidatable
-    {
-        TValue Create(TParam1 param1, TParam2 param2);
-    }
-
     public class FactoryTyped<TParam1, TParam2, TValue> : IFactoryTyped<TParam1, TParam2, TValue>
     {
         [Inject]
@@ -74,12 +54,6 @@ namespace ModestTree.Zenject
             return _container.ValidateObjectGraph<TValue>(
                 typeof(TParam1), typeof(TParam2));
         }
-    }
-
-    // Three parameters
-    public interface IFactoryTyped<TParam1, TParam2, TParam3, TValue> : IValidatable
-    {
-        TValue Create(TParam1 param1, TParam2 param2, TParam3 param3);
     }
 
     public class FactoryTyped<TParam1, TParam2, TParam3, TValue>
@@ -98,12 +72,6 @@ namespace ModestTree.Zenject
             return _container.ValidateObjectGraph<TValue>(
                 typeof(TParam1), typeof(TParam2), typeof(TParam3));
         }
-    }
-
-    // Four parameters
-    public interface IFactoryTyped<TParam1, TParam2, TParam3, TParam4, TValue> : IValidatable
-    {
-        TValue Create(TParam1 param1, TParam2 param2, TParam3 param3, TParam4 param4);
     }
 
     public class FactoryTyped<TParam1, TParam2, TParam3, TParam4, TValue>
