+++ UnityProject/Assets/Zenject/Main/Scripts/Factories/GameObjectInstantiator.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Collections.Generic;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
 using UnityEngine;
 
 namespace ModestTree.Zenject
