+++ UnityProject/Assets/Zenject/Main/Scripts/Factories/IValidatable.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using ModestTree.Zenject.Api.Exceptions;
+
+namespace ModestTree.Zenject.Api.Factories
+{
+    public interface IValidatable
+    {
+        IEnumerable<ZenjectResolveException> Validate();
+    }
+}
