+++ UnityProject/Assets/Zenject/Main/Scripts/Main/DiContainer.cs
@@ -1,9 +1,8 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using System.Text;
+using ModestTree.Zenject.Api.Exceptions;
+using ModestTree.Zenject.Api.Factories;
 using UnityEngine;
 
 namespace ModestTree.Zenject
