+++ UnityProject/Assets/Zenject/Main/Scripts/Misc/InitializablePrioritiesInstaller.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-using ModestTree.Zenject;
+using ModestTree.Zenject.Api.Misc;
 
 namespace ModestTree.Zenject
 {
