+++ UnityProject/Assets/Zenject/Main/Scripts/Misc/KernelUtil.cs
@@ -1,4 +1,6 @@
 using System;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Misc;
 
 namespace ModestTree.Zenject
 {
