+++ UnityProject/Assets/Zenject/Main/Scripts/Misc/StandardKernel.cs
@@ -1,7 +1,8 @@
 using System;
 using System.Collections.Generic;
-using UnityEngine;
+using ModestTree.Zenject.Api;
 using System.Linq;
+using ModestTree.Zenject.Api.Misc;
 
 namespace ModestTree.Zenject
 {
