+++ UnityProject/Assets/Zenject/Main/Scripts/Misc/StandardUnityInstaller.cs
@@ -1,5 +1,7 @@
 using System;
 using System.Linq;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Misc;
 
 namespace ModestTree.Zenject
 {
