+++ UnityProject/Assets/Zenject/Main/Scripts/Providers/GameObjectSingletonProviderFromPrefab.cs
@@ -1,6 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+using ModestTree.Zenject.Api.Factories;
 using UnityEngine;
 
 namespace ModestTree.Zenject
