+++ UnityProject/Assets/Zenject/Main/Scripts/Providers/ProviderBase.cs
@@ -1,5 +1,8 @@
 using System;
 using System.Collections.Generic;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+
 namespace ModestTree.Zenject
 {
     public abstract class ProviderBase : IDisposable
