+++ UnityProject/Assets/Zenject/Main/Scripts/Providers/TransientProvider.cs
@@ -1,6 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using ModestTree.Zenject.Api;
+using ModestTree.Zenject.Api.Exceptions;
+
 namespace ModestTree.Zenject
 {
     public class TransientProvider : ProviderBase
