+++ UnityProject/Assets/Zenject/Main/Scripts/Zenject.csproj
@@ -7,7 +7,7 @@
     <ProjectGuid>{46F25A62-2E29-48CB-95F3-BDBCB0976DDC}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Zenject</RootNamespace>
+    <RootNamespace>ModestTree.Zenject</RootNamespace>
     <AssemblyName>Zenject</AssemblyName>
     <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
@@ -58,14 +58,12 @@
     <Compile Include="Binders\InjectContext.cs" />
     <Compile Include="Binders\ReferenceBinder.cs" />
     <Compile Include="Binders\ValueBinder.cs" />
-    <Compile Include="Exceptions\ZenjectException.cs" />
     <Compile Include="Factories\Factory.cs" />
     <Compile Include="Factories\FactoryCustom.cs" />
     <Compile Include="Factories\FactoryMethod.cs" />
     <Compile Include="Factories\FactoryTyped.cs" />
     <Compile Include="Factories\GameObjectFactory.cs" />
     <Compile Include="Factories\GameObjectInstantiator.cs" />
-    <Compile Include="Factories\IFactory.cs" />
     <Compile Include="Factories\Instantiator.cs" />
     <Compile Include="Factories\KeyedFactory.cs" />
     <Compile Include="Factories\ListFactory.cs" />
@@ -76,19 +74,19 @@
     <Compile Include="Main\BindingValidator.cs" />
     <Compile Include="Main\CompositionRoot.cs" />
     <Compile Include="Main\DiContainer.cs" />
-    <Compile Include="Main\IInstaller.cs" />
+    <Compile Include="Main\IInstaller.cs">
+      <SubType>Code</SubType>
+    </Compile>
     <Compile Include="Main\Installer.cs" />
     <Compile Include="Main\MonoInstaller.cs" />
     <Compile Include="Misc\DependencyRootStandard.cs" />
     <Compile Include="Misc\DisposablePrioritiesInstaller.cs" />
     <Compile Include="Misc\DisposablesHandler.cs" />
     <Compile Include="Misc\IDependencyRoot.cs" />
-    <Compile Include="Misc\IInitializable.cs" />
     <Compile Include="Misc\IKernel.cs" />
     <Compile Include="Misc\InitializableHandler.cs" />
     <Compile Include="Misc\InitializablePrioritiesInstaller.cs" />
     <Compile Include="Misc\InjectAttribute.cs" />
-    <Compile Include="Misc\ITickable.cs" />
     <Compile Include="Misc\KernelUtil.cs" />
     <Compile Include="Misc\LookupInProgressAdder.cs" />
     <Compile Include="Misc\StandardKernel.cs" />
@@ -115,7 +113,12 @@
     <Compile Include="Util\TypeExtensions.cs" />
     <Compile Include="Util\UnityUtil.cs" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <ProjectReference Include="Zenject.Api.csproj">
+      <Project>{73a60b22-4135-4b72-9306-93263aaa9c2b}</Project>
+      <Name>Zenject.Api</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
        Other similar extension points exist, see Microsoft.Common.targets.
